wip getting phoenix socket subscription working
