+++ assets/js/util/apolloClient.js
@@ -1,0 +1,28 @@
+
+// GraphQL
+import { ApolloClient } from 'apollo-client';
+import { createHttpLink } from 'apollo-link-http';
+import { setContext } from 'apollo-link-context';
+import { InMemoryCache } from 'apollo-cache-inmemory';
+import { store } from '../store/configureStore';
+
+const httpLink = createHttpLink({
+  uri: "/graphql"
+})
+
+const authLink = setContext((_, { headers }) => {
+  const token = store.getState().auth.apikey
+  return {
+    headers: {
+      ...headers,
+      authorization: token ? `Bearer ${token}` : ""
+    }
+  }
+})
+
+const apolloClient = new ApolloClient({
+  link: authLink.concat(httpLink),
+  cache: new InMemoryCache(),
+})
+
+export default apolloClient
