+++ assets/package.json
@@ -8,6 +8,7 @@
   "dependencies": {
     "@material-ui/core": "^1.0.0",
     "@material-ui/icons": "^1.0.0",
+    "@absinthe/socket-apollo-link": "^0.1.11",
     "apollo-client-preset": "^1.0.8",
     "apollo-link-context": "^1.0.8",
     "axios": "^0.18.0",
